+++ docs/ngdocs/js/docs-setup.js
@@ -1,0 +1,157 @@
+NG_DOCS={
+  "sections": {
+    "api": "AngularJS Documentation"
+  },
+  "pages": [
+    {
+      "section": "api\nSets the $pw service object into the local scope",
+      "id": "pwGlobals",
+      "shortName": "pwGlobals",
+      "type": "directive",
+      "moduleName": "pwGlobals",
+      "keywords": "$pw api bind directive expression globals json local object postworld pw pw-globals pwglobals scope service sets"
+    },
+    {
+      "section": "api",
+      "id": "postworld",
+      "shortName": "postworld",
+      "type": "overview",
+      "moduleName": "postworld",
+      "shortDescription": "The core module for working with Postworld on the front-end.",
+      "keywords": "api core front-end module overview postworld working"
+    },
+    {
+      "section": "api",
+      "id": "postworld.$_",
+      "shortName": "postworld.$_",
+      "type": "service",
+      "moduleName": "postworld",
+      "keywords": "$_ api boolean converts false method postworld rename representation service string strings stringtoboolean true"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwAutofocus",
+      "shortName": "pwAutofocus",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Automatically focuses the input element it&#39;s applied to.",
+      "keywords": "api applied automatically directive element focuses input postworld"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwBackgroundImage",
+      "shortName": "pwBackgroundImage",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Adds a background image style property to an element",
+      "keywords": "adds api background binding directive element image large post postworld property pw-background-image pwbackgroundimage sizes style url"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwClickPreventDefault",
+      "shortName": "pwClickPreventDefault",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Prevents the default click action on the element.",
+      "keywords": "action api click default directive element postworld prevents"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwClickStopPropagation",
+      "shortName": "pwClickStopPropagation",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Stops the click event from propagating beyond the given element.",
+      "keywords": "api click directive element event postworld propagating stops"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwEval",
+      "shortName": "pwEval",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Evaluates a string as javascript at the time of loading.",
+      "keywords": "api change changes context default directive evalcontext evaltimeout evaluate evaluates evaluating evalwatch expression initializing javascript libraries loading milliseonds optional options postworld pweval re-evalute scope string third-party time timeout watch window works"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwFeed",
+      "shortName": "pwFeed",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Generates a feed from predefined feed paramters within the specified element.",
+      "keywords": "ability api current directive element feed feedid feeds generates implimentations list live-feed load_increment newsfeed non-scrolling offset order_by paramters post post_status post_type posts_per_page postworld predefined preload publish pw pwfeed query rename view"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwGlobals",
+      "shortName": "pwGlobals",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Sets the $pw service object into the local scope.",
+      "keywords": "$pw api bind directive expression globals json local object postworld pw pw-globals pwglobals scope service sets"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwHoverClass",
+      "shortName": "pwHoverClass",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Adds specified class(es) to an element on mouseover,",
+      "keywords": "add adds api class classes directive element hovered hoveroffdelay hoverondelay milliseconds mouseleave mouseover optional postworld pwhoverclass removed removes"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwInclude",
+      "shortName": "pwInclude",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Used to include a Postworld template partial within an isolated scope.",
+      "keywords": "$scope api assign assigned class data directive dynamic easily element enable galleries html include include-post includeclass included includeenable includemeta includepost includevars instances isolated load loading meta object panels partial post postworld prevent psuedo-path pw-include pwinclude scope template templates vars"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwLoadPost",
+      "shortName": "pwLoadPost",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Loads a post template and injects it with data.",
+      "keywords": "add api assign boolean classes data derive directive element expression injects loading loads optional post postclass postid postloading postquery postview postworld query registered template vars view"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwPost",
+      "shortName": "pwPost",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Provides service functions to post templates.",
+      "keywords": "api directive functions post postworld service templates"
+    },
+    {
+      "section": "api",
+      "id": "postworld.directive:pwShareLink",
+      "shortName": "pwShareLink",
+      "type": "directive",
+      "moduleName": "postworld",
+      "shortDescription": "Generates a share link for a particular post ID.",
+      "keywords": "api directive expression generate generates link map post postworld pw-share-link pwsharelink required sets share share-link-post-id sharelink sharelinkdynamic sharelinkpostid true watch"
+    },
+    {
+      "section": "api",
+      "id": "postworld.pwData",
+      "shortName": "postworld.pwData",
+      "type": "service",
+      "moduleName": "postworld",
+      "keywords": "$pwdata $q $resource ajax api args arguments called calls data database doesn easy example exist failure false fname function functions gettemplate hello key list meta meta_key method optional panels php post post_type postworld pwdata rename requested retreives returns server service sets setwpusermeta simplified sub_key subdir subkey success template type url user user_id userdata values view wordpress wpajax wrapper"
+    }
+  ],
+  "apis": {
+    "api": true
+  },
+  "html5Mode": false,
+  "editExample": true,
+  "startPage": "#/api",
+  "scripts": [
+    "angular.min.js"
+  ]
+};